+++ src/app/[id]/page.tsx
@@ -1,40 +1,46 @@
-"use client";
-
-import { use } from "react";
-
+import { Suspense } from "react";
 import { Skeleton } from "@/components/ui/skeleton";
-import { useBook } from "./_hooks/use-book";
 import { PDFBook } from "./_components/pdf-book";
+import { db } from "@/server/db";
+import { notFound } from "next/navigation";
 
-export default function Page(props: { params: Promise<{ id: string }> }) {
-  const params = use(props.params);
-  const { loading, book, user, history, loadHistory } = useBook(params.id);
+export default async function Page({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  const { id } = await params;
 
-  if (loading) {
-    return (
-      <div className="my-6 overflow-hidden">
-        <div className="mb-6">
-          <Skeleton className="my-2 h-10 w-full" />
-          <hr className="bg-primary h-[6px] max-w-[36px] border-0" />
-        </div>
-        <Skeleton className="h-[600px]" />
-      </div>
-    );
-  }
   return (
-    book && (
-      <div className="my-6 overflow-hidden">
-        <div className="mb-6">
-          <h1 className="my-2 text-3xl font-bold md:text-4xl">{book.title}</h1>
-          <hr className="bg-primary h-[6px] max-w-[36px] border-0" />
-        </div>
-        <PDFBook
-          book={book}
-          user={user}
-          history={history}
-          loadHistory={loadHistory}
-        />
+    <Suspense
+      fallback={
+        <>
+          <div className="mb-6">
+            <Skeleton className="my-2 h-10 w-full" />
+            <hr className="bg-primary h-[6px] max-w-[36px] border-0" />
+          </div>
+          <Skeleton className="h-[600px]" />
+        </>
+      }
+    >
+      <SuspenseBoundary id={id} />
+    </Suspense>
+  );
+}
+
+async function SuspenseBoundary({ id }: { id: string }) {
+  const book = await db.book.findUnique({
+    where: { id },
+  });
+
+  if (!book) notFound();
+  return (
+    <>
+      <div className="mb-6">
+        <h1 className="my-2 text-3xl font-bold md:text-4xl">{book.title}</h1>
+        <hr className="bg-primary h-[6px] max-w-[36px] border-0" />
       </div>
-    )
+      <PDFBook book={book} />
+    </>
   );
 }
