--- src/app/api/me/route.tsx
@@ -1,7 +1,0 @@
-import { auth } from "@/server/auth";
-
-export async function GET() {
-  const session = await auth();
-
-  return Response.json({ user: session?.user });
-}
